Introduce HadoopFsRelationOptimizer  (#976)
This is a minor refactor:

- Introduce a new helper object named HadoopFsRelationOptimizer, this helper use to replace FileFormat to Optimized FileFormat, code move from FileSourceStrategy .

- Revert change of FileSourceStrategy except call HadoopFsRelationOptimizer.tryOptimize method
